+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Fix `Sevencop/OrderField` on dstr.
+
 ## 0.4.0 - 2022-06-18
 
 ### Added